+++ packages/file-association/src/ObjectManager/DefaultObjectManager.php
@@ -17,6 +17,8 @@ use Rekalogika\File\Association\Contracts\ClassMetadataFactoryInterface;
 use Rekalogika\File\Association\Contracts\FilePropertyManagerInterface;
 use Rekalogika\File\Association\Contracts\ObjectIdResolverInterface;
 use Rekalogika\File\Association\Contracts\ObjectManagerInterface;
+use Rekalogika\File\Association\Model\ObjectOperationResult;
+use Rekalogika\File\Association\Model\ObjectOperationType;
 use Rekalogika\File\Association\Util\ProxyUtil;
 
 final readonly class DefaultObjectManager implements ObjectManagerInterface
@@ -39,56 +41,80 @@ final readonly class DefaultObjectManager implements ObjectManagerInterface
      * Called when the object is saved
      */
     #[\Override]
-    public function flushObject(object $object): void
+    public function flushObject(object $object): ObjectOperationResult
     {
         $class = $this->getClass($object);
         $classMetadata = $this->classMetadataFactory->getClassMetadata($class);
         $id = $this->objectIdResolver->getObjectId($object);
+        $propertyResults = [];
 
         foreach ($classMetadata->getProperties() as $propertyMetadata) {
-            $this->filePropertyManager->flushProperty(
+            $propertyResults[] = $this->filePropertyManager->flushProperty(
                 propertyMetadata: $propertyMetadata,
                 object: $object,
                 id: $id,
             );
         }
+
+        return new ObjectOperationResult(
+            type: ObjectOperationType::Flush,
+            class: $class,
+            objectId: $id,
+            propertyResults: $propertyResults,
+        );
     }
 
     /**
      * Called when the object is removed
      */
     #[\Override]
-    public function removeObject(object $object): void
+    public function removeObject(object $object): ObjectOperationResult
     {
         $class = $this->getClass($object);
         $classMetadata = $this->classMetadataFactory->getClassMetadata($class);
         $id = $this->objectIdResolver->getObjectId($object);
+        $propertyResults = [];
 
         foreach ($classMetadata->getProperties() as $propertyMetadata) {
-            $this->filePropertyManager->removeProperty(
+            $propertyResults[] = $this->filePropertyManager->removeProperty(
                 propertyMetadata: $propertyMetadata,
                 object: $object,
                 id: $id,
             );
         }
+
+        return new ObjectOperationResult(
+            type: ObjectOperationType::Remove,
+            class: $class,
+            objectId: $id,
+            propertyResults: $propertyResults,
+        );
     }
 
     /**
      * Called after the object is loaded
      */
     #[\Override]
-    public function loadObject(object $object): void
+    public function loadObject(object $object): ObjectOperationResult
     {
         $class = $this->getClass($object);
         $classMetadata = $this->classMetadataFactory->getClassMetadata($class);
         $id = $this->objectIdResolver->getObjectId($object);
+        $propertyResults = [];
 
         foreach ($classMetadata->getProperties() as $propertyMetadata) {
-            $this->filePropertyManager->loadProperty(
+            $propertyResults[] = $this->filePropertyManager->loadProperty(
                 propertyMetadata: $propertyMetadata,
                 object: $object,
                 id: $id,
             );
         }
+
+        return new ObjectOperationResult(
+            type: ObjectOperationType::Load,
+            class: $class,
+            objectId: $id,
+            propertyResults: $propertyResults,
+        );
     }
 }